Put BaseConfig in autoload, in "constants.php"
