+++ tests/phpstan_bootstrap.php
@@ -35,5 +35,3 @@
 $root = dirname(__DIR__);
 
 require $root . '/vendor/autoload.php';
-
-(new \Glpi\Config\LegacyConfigurators\BaseConfig($root, 'development'))->execute();
